move function to BankAccounts.js
